+++ CHANGELOG.md
@@ -6,6 +6,8 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.3.2] - 2026-09-09
+
 ### Added
 - `TimeSeries.batch` and `T0Forecaster.predict_from_time_series` for integrations
   that construct complete T0 inputs, including known-future covariates.